requirements.txt
Flask
tensorflow
numpy
pandas
scikit-learn
gunicorn
Pillow